+++ server/data/user/src/main/java/ru/kima/sonar/server/data/user/model/portfolio/PortfolioEntry.kt
@@ -1,13 +1,13 @@
 package ru.kima.sonar.server.data.user.model.portfolio
 
-import java.math.BigInteger
+import java.math.BigDecimal
 
 data class PortfolioEntry(
     val id: Long,
     val portfolioId: Long,
     val securityUid: String,
     val name: String,
-    val lowPrice: BigInteger,
-    val highPrice: BigInteger,
+    val lowPrice: BigDecimal,
+    val highPrice: BigDecimal,
     val note: String,
 )
